backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-admin-logs>=1.0.0
django-otp>=1.3.0
mssql-django>=1.3
pyodbc>=4.0.39
python-dotenv>=1.0.0
msal>=1.25.0
msgraph-core>=0.2.2
azure-identity>=1.15.0
aiohttp>=3.9.0 